move getCurrentUser
